fix: adding of child nodes to parent node
The order of parsed classes is undefined, that is why the child class
cannot be added to the parent class because the parent class can be not
defined in the map of classes.

The routine of adding child classes to the parent class is moved to the
moment when all files are parsed.