new: automatically generate short description for external classes
Previously the property "externalLinks" were used by "inheritance-diagram" library only to directly point to the external URL.
Now this property is under the responsibility of this plugin. Based on it short additional documentation for external classes is generating. This gives the way to display the inheritance diagram for that external class used in some project and the link to that class will be available under "see" section.