fix: References to the classes that are under some namespace.
For example we have "ns" as a namespace. That namespace has definition for the class "A" and class "B", that extends class "A".
Before the output diagram for class B shows the parent class, but without the link to the class "A", because it is actually placed in the namespace. The link to class "A" was something like ".../A.html", but should be ".../ns.A.html". The using of long name resolves this.
As a side effect, it is also good to see the fully qualified name on the diagram in order to avoid possible confusion if few namespaces will contain the same class names.

fixes #2